add find nearby places (with category) & change code some things
