+++ PlacesNearMe/app/src/main/java/com/example/placesnearme/Model/DiaDiem.java
@@ -12,10 +12,13 @@ public class DiaDiem implements Serializable {
     GeoPoint location;
     String madiadiem;
     String tendiadiem;
+    Reviews reviews;
+
 
     public DiaDiem() { }
 
-    public DiaDiem(List<String> danhmuc, String diachi, List<String> hinhAnh, GeoPoint location, String madiadiem, String tendiadiem) {
+    public DiaDiem(List<String> danhmuc, String diachi, List<String> hinhAnh, GeoPoint location,
+                   String madiadiem, String tendiadiem) {
         this.danhmuc = danhmuc;
         this.diachi = diachi;
         this.hinhAnh = hinhAnh;
