+++ PlacesNearMe/app/src/main/res/layout/activity_result_list_and_map.xml
@@ -1,14 +1,20 @@
 <?xml version="1.0" encoding="utf-8"?>
 <RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".View.ResultListAndMapActivity">
+    android:layout_height="match_parent">
+
+    <fragment
+        android:id="@+id/map"
+        android:name="com.google.android.gms.maps.SupportMapFragment"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent" />
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/listDiaDiem"
         android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:padding="2dp"/>
+        android:layout_height="match_parent"
+        android:padding="2dp"
+        android:background="@color/ratingLightGray"
+        android:visibility="gone"/>
 </RelativeLayout>
