+++ PlacesNearMe/app/src/main/res/layout/list_item_places.xml
@@ -7,7 +7,7 @@
     android:layout_width="match_parent"
     android:layout_height="wrap_content"
     app:cardElevation="10dp"
-    app:cardCornerRadius="10dp"
+    app:cardCornerRadius="5dp"
     android:layout_margin="5dp"
     app:cardBackgroundColor="@android:color/holo_blue_dark">
 
@@ -18,8 +18,6 @@
         android:layout_gravity="center"
         android:layout_marginLeft="5dp"
         android:layout_marginRight="5dp"
-        android:layout_marginTop="5dp"
-        android:layout_marginBottom="5dp"
         android:background="@android:color/white">
 
         <LinearLayout
@@ -36,14 +34,23 @@
                 android:src="@mipmap/ic_launcher" />
 
             <TextView
-                android:id="@+id/txtDuration"
+                android:id="@+id/txtCountReview"
                 android:layout_width="wrap_content"
                 android:layout_height="wrap_content"
-                android:text="5 mins"
+                android:text="5 people reviews"
                 android:textColor="@android:color/darker_gray"
                 android:textSize="12sp"
                 android:layout_gravity="center|bottom"
                 android:layout_marginTop="10dp" />
+
+            <RatingBar
+                android:id="@+id/ratingBar"
+                android:layout_width="wrap_content"
+                android:layout_height="wrap_content"
+                android:layout_gravity="center|bottom"
+                style="?android:attr/ratingBarStyleSmall"
+                android:rating="4"
+                android:theme="@style/RatingBar"/>
         </LinearLayout>
 
         <View
