+++ PlacesNearMe/app/src/main/res/menu/headermenu.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:id="@+id/header_menu_filter"
+        android:icon="@drawable/ic_filter_list_black_24dp"
+        android:title="Filter"
+        app:showAsAction="always" />
+
+    <item
+        android:icon="@drawable/ic_more_vert_black_24dp"
+        android:title=""
+        app:showAsAction="always">
+
+        <menu>
+            <item
+                android:id="@+id/header_menu_map"
+                android:icon="@drawable/ic_map_black_24dp"
+                android:title="View by map"
+                app:showAsAction="ifRoom" />
+            <item
+                android:id="@+id/header_menu_list"
+                android:icon="@drawable/ic_list"
+                android:title="View by list"
+                app:showAsAction="ifRoom" />
+        </menu>
+    </item>
+</menu>
