+++ PlacesNearMe/app/src/main/res/values/styles.xml
@@ -7,4 +7,9 @@
         <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
         <item name="colorAccent">@color/colorAccent</item>
     </style>
+
+    <style name="RatingBar" parent="Theme.AppCompat">
+        <item name="colorControlNormal">@color/ratingLightGray</item>
+        <item name="colorControlActivated">@color/ratingYellow</item>
+    </style>
 </resources>
